refactor(hypav3): move handleDualAction to utils module
