+++ src/lib/Others/HypaV3Modal.svelte
@@ -29,7 +29,7 @@
     SearchUI,
     SearchResult,
   } from "./HypaV3Modal/types";
-  import { alertConfirmTwice } from "./HypaV3Modal/utils";
+  import { alertConfirmTwice, handleDualAction } from "./HypaV3Modal/utils";
 
   const hypaV3DataState = $derived(
     DBState.db.characters[$selectedCharID].chats[
@@ -675,73 +675,6 @@
         error: `Error occurred: ${error.message}`,
       };
     }
-  }
-
-  type DualActionParams = {
-    onMainAction?: () => void;
-    onAlternativeAction?: () => void;
-  };
-
-  function handleDualAction(node: HTMLElement, params: DualActionParams = {}) {
-    const DOUBLE_TAP_DELAY = 300;
-
-    const state = {
-      lastTap: 0,
-      tapTimeout: null,
-    };
-
-    const handleTouch = (event: TouchEvent) => {
-      const currentTime = new Date().getTime();
-      const tapLength = currentTime - state.lastTap;
-
-      if (tapLength < DOUBLE_TAP_DELAY && tapLength > 0) {
-        // Double tap detected
-        event.preventDefault();
-        window.clearTimeout(state.tapTimeout); // Cancel the first tap timeout
-        params.onAlternativeAction?.();
-        state.lastTap = 0; // Reset state
-      } else {
-        state.lastTap = currentTime; // First tap
-        // Delayed single tap execution
-        state.tapTimeout = window.setTimeout(() => {
-          if (state.lastTap === currentTime) {
-            // If no double tap occurred
-            params.onMainAction?.();
-          }
-        }, DOUBLE_TAP_DELAY);
-      }
-    };
-
-    const handleClick = (event: MouseEvent) => {
-      if (event.shiftKey) {
-        params.onAlternativeAction?.();
-      } else {
-        params.onMainAction?.();
-      }
-    };
-
-    if ("ontouchend" in window) {
-      // Mobile environment
-      node.addEventListener("touchend", handleTouch);
-    } else {
-      // Desktop environment
-      node.addEventListener("click", handleClick);
-    }
-
-    return {
-      destroy() {
-        if ("ontouchend" in window) {
-          node.removeEventListener("touchend", handleTouch);
-        } else {
-          node.removeEventListener("click", handleClick);
-        }
-
-        window.clearTimeout(state.tapTimeout); // Cleanup timeout
-      },
-      update(newParams: DualActionParams) {
-        params = newParams;
-      },
-    };
   }
 </script>
